use unordered_map instead of TypeUnorderedMapContainer template collection for in-map object storing v0.5. DANGEROWS! NEED TEST and measurements CPU load and resource leaks! related to #1165, #1083, #852

+++ src/game/Unit.cpp
@@ -11215,7 +11215,7 @@ uint32 Unit::GetCreatePowers( Powers power ) const
 
 void Unit::AddToWorld()
 {
-    Object::AddToWorld();
+    WorldObject::AddToWorld();
     ScheduleAINotify(0);
 }
 
@@ -11238,11 +11238,14 @@ void Unit::RemoveFromWorld()
         GetViewPoint().Event_RemovedFromWorld();
     }
 
-    Object::RemoveFromWorld();
+    WorldObject::RemoveFromWorld();
 }
 
-void Unit::CleanupsBeforeDelete()
+void Unit::CleanupsBeforeDelete(bool force)
 {
+    if (!IsInWorld() && !force)
+        return;
+
     if (m_uint32Values)                                      // only for fully created object
     {
         if (GetVehicle())
@@ -11251,18 +11254,18 @@ void Unit::CleanupsBeforeDelete()
             RemoveVehicleKit();
         InterruptNonMeleeSpells(true);
         KillAllEvents(false);                      // non-delatable (currently casted spells) will not deleted now but it will deleted at call in Map::RemoveAllObjectsInRemoveList
-        if (IsInWorld())
-            CombatStop();
+        CombatStop();
         ClearComboPointHolders();
-        DeleteThreatList();
-        if (GetTypeId()==TYPEID_PLAYER)
+        if (CanHaveThreatList())
+            DeleteThreatList();
+        if (GetTypeId() == TYPEID_PLAYER)
             getHostileRefManager().setOnlineOfflineState(false);
         else
             getHostileRefManager().deleteReferences();
         RemoveAllAuras(AURA_REMOVE_BY_DELETE);
         GetUnitStateMgr().InitDefaults(false);
     }
-    WorldObject::CleanupsBeforeDelete();
+    WorldObject::CleanupsBeforeDelete(force);
 }
 
 CharmInfo* Unit::InitCharmInfo(Unit *charm)
